requirements.txt
streamlit
PyMuPDF
sentence_transformers
pinecone-client